AddDatabase

+++ ManicOceanic.DOMAIN/Entities/Products/Product.cs
@@ -1,14 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace ManicOceanic.DOMAIN.Entities.Products
 {
-  public class Product
+    public class Product
   {
     public Guid Id { get; set; }
-    public Int64 ProductNumber { get; set; }
+    public int ProductNumber { get; set; }
     public string Name { get; set; }
     public decimal Price { get; set; }
     public string ImageUrl { get; set; }
